Update UPA to v2.0.3
