+++ core/src/submitSolutionsFromFile.ts
@@ -1,14 +1,14 @@
 import { loadDemoAppInstance, demoAppInstance } from "./utils";
-import { OffChainSubmission } from "@nebrazkp/upa/sdk";
 import * as ethers from "ethers";
 import { command, option, number, string } from "cmd-ts";
 import { options, config } from "@nebrazkp/upa/tool";
-import { DemoApp, DemoApp__factory } from "../typechain-types";
+import { DemoApp__factory } from "../typechain-types";
+import { submitSolution } from "./utils";
 
 export const submitSolutionsFromFile = command({
   name: "submit-solutions-from-file",
   args: {
-    endpoint: options.endpoint(),
+    chainEndpoint: options.chainEndpoint(),
     keyfile: options.keyfile(),
     password: options.password(),
     maxFeePerGasGwei: options.maxFeePerGasGwei(),
@@ -29,13 +29,13 @@ export const submitSolutionsFromFile = command({
     "Submit solutions to demo-app contract from a file. \
     Proofs must be verified by UPA first.",
   handler: async function ({
-    endpoint,
+    chainEndpoint,
     keyfile,
     password,
     demoAppInstanceFile,
     submissionFile,
   }): Promise<void> {
-    const provider = new ethers.JsonRpcProvider(endpoint);
+    const provider = new ethers.JsonRpcProvider(chainEndpoint);
     const wallet = await config.loadWallet(keyfile, password, provider);
     const demoAppInstance = loadDemoAppInstance(demoAppInstanceFile);
     const demoApp = DemoApp__factory.connect(demoAppInstance.demoApp).connect(
@@ -86,43 +86,3 @@ export const submitSolutionsFromFile = command({
     });
   },
 });
-
-/**
- * Submits a solution to the demo app contract. The solution
- * should already have been verified by the UPA contract.
- *
- * @param wallet - The wallet used to sign the transaction.
- * @param demoApp - The instance of the demo app contract.
- * @param nonce - The nonce value for the transaction.
- * @param submission - The submission object containing the solution.
- * @param solutionIdx - The index of the solution in the submission.
- * @returns A promise that resolves to the transaction response.
- */
-export async function submitSolution(
-  _wallet: ethers.AbstractSigner,
-  demoApp: DemoApp,
-  nonce: number,
-  submission: OffChainSubmission,
-  solutionIdx: number
-): Promise<ethers.ContractTransactionResponse> {
-  const solution = submission.inputs[solutionIdx];
-  console.log(`Submitted solution ${solution}`);
-  if (submission.isMultiProofSubmission()) {
-    // If the proof was part of a multi-proof submission, we
-    // need to pass a proof reference to the demo-app
-    // contract so it can check the proof's verification
-    // status.
-    return demoApp.submitSolutionWithProofReference(
-      solution,
-      submission.computeProofReference(solutionIdx)!.solidity(),
-      { nonce: nonce }
-    );
-  } else {
-    // If the proof was sent in a single-proof submission, we
-    // only need to pass the solution. A proof reference is not
-    // necessary.
-    return demoApp.submitSolution(solution, {
-      nonce: nonce,
-    });
-  }
-}
