+++ frontend/package.json
@@ -10,7 +10,7 @@
     "format": "prettier . --write"
   },
   "dependencies": {
-    "@nebrazkp/upa": "../upa/package.tgz",
+    "@nebrazkp/upa": "2.0.3",
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18"
